backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
pydantic-settings
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
celery==5.3.6
redis==5.0.3
boto3==1.34.74
beautifulsoup4==4.12.3
requests==2.31.0
python-multipart
pypdf
twilio==8.13.0
sendgrid==6.11.0
websockets==12.0
httpx==0.27.0
python-dotenv==1.0.1
aiofiles==23.2.1
email-validator==2.1.1
langchain-google-genai
langchain-groq
langchain-openai
langchain-core
psycopg2-binary
numpy
scikit-learn
scipy
joblib
praw==7.7.1
tweepy==4.14.0
deep-translator
